use the 64bit types for id in the admin functions

+++ src/source.h
@@ -105,7 +105,7 @@ void source_clear_listeners (source_t *source);
 void source_clear_source (source_t *source);
 source_t *source_find_mount(const char *mount);
 source_t *source_find_mount_raw(const char *mount);
-client_t *source_find_client(source_t *source, int id);
+client_t *source_find_client(source_t *source, uint64_t id);
 int source_compare_sources(void *arg, void *a, void *b);
 void source_free_source(source_t *source);
 void source_main(source_t *source);
